dashboard: move to hash router
