+++ dashboard/frontend/src/services.js
@@ -1,11 +1,4 @@
-function getHost() {
-  const prefix = "/tfjobs";
-  const pLen = prefix.length;
-
-  const fullPath = window.location.pathname;
-  const keepLen = fullPath.indexOf(prefix) + pLen;
-  return fullPath.substr(0, keepLen);
-}
+import { getHost } from "./utils";
 
 const host = getHost();
